modify validateSchema struct
