+++ src/schemaValidation/index.ts
@@ -1,1 +1,3 @@
-export * from "././validateSchema.js";
+export { validateComposeSchema } from "./validateComposeSchema.js";
+export { validateManifestSchema } from "./validateManifestSchema.js";
+export { validateSetupWizardSchema } from "./validateSetupWizardSchema.js";
