+++ src/schemaValidation/utils.ts
@@ -1,0 +1,65 @@
+import { ErrorObject } from "ajv";
+
+// validate.errors = [
+//   {
+//     keyword: "pattern",
+//     dataPath: "/avatar",
+//     schemaPath: "#/properties/avatar/pattern",
+//     params: { pattern: "^/(ipfs|bzz)/w+$" },
+//     message: 'should match pattern "^/(ipfs|bzz)/w+$"'
+//   },
+//   {
+//     keyword: "pattern",
+//     dataPath: "/image/hash",
+//     schemaPath: "#/properties/image/properties/hash/pattern",
+//     params: { pattern: "^/(ipfs|bzz)/w+$" },
+//     message: 'should match pattern "^/(ipfs|bzz)/w+$"'
+//   },
+//   {
+//     keyword: "type",
+//     dataPath: "/image/size",
+//     schemaPath: "#/properties/image/properties/size/type",
+//     params: { type: "number" },
+//     message: "should be number"
+//   },
+//   {
+//     keyword: "pattern",
+//     dataPath: "/image/size",
+//     schemaPath: "#/properties/image/properties/size/pattern",
+//     params: { pattern: "^d+$" },
+//     message: 'should match pattern "^d+$"'
+//   },
+//   {
+//     keyword: "required",
+//     dataPath: "",
+//     schemaPath: "#/required",
+//     params: { missingProperty: "license" },
+//     message: "should have required property 'license'"
+//   }
+// ];
+
+/**
+ *
+ * @param errorObject from AJV:
+ * {
+ *   keyword: "pattern",
+ *   dataPath: "/avatar",
+ *   schemaPath: "#/properties/avatar/pattern",
+ *   params: { pattern: "^/(ipfs|bzz)/w+$" },
+ *   message: 'should match pattern "^/(ipfs|bzz)/w+$"'
+ * }
+ * @returns errorMessage:
+ * "manifest.avatar should match pattern "^/(ipfs|bzz)/w+$""
+ */
+export function processError(
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  errorObject: ErrorObject<string, Record<string, any>, unknown>,
+  releaseFileType: "compose" | "manifest" | "setupWizard"
+): string {
+  const { schemaPath, message } = errorObject;
+  const path = `${releaseFileType}${schemaPath}`.replace(
+    new RegExp("/", "g"),
+    "."
+  );
+  return `${path} ${message}`;
+}
